aivideochat/requirements.txt
asgiref==3.8.1
channels==4.2.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
sqlparse==0.5.3
